requirements.txt
# Kuber - Algorithmic Trading Platform
# Core Dependencies

# Robinhood API
robin_stocks>=3.4.0
pyotp>=2.9.0

# Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Technical Analysis
ta
# Data Download
yfinance>=0.2.36
polygon-api-client>=1.12.0

# Machine Learning & RL
scikit-learn>=1.3.0
torch>=2.3.0
stable-baselines3>=2.2.1
gymnasium>=0.29.1
shimmy>=1.3.0

# Backtesting
vectorbt>=0.25.0

# Visualization
plotly>=5.18.0
streamlit>=1.29.0
matplotlib>=3.8.0

# Database & Storage
sqlalchemy>=2.0.0

# Scheduling & Background Tasks
schedule>=1.2.0
python-dotenv>=1.0.0

# Utilities
rich>=13.6.0
tabulate>=0.9.0
tqdm>=4.66.0

# HTTP & Async
requests>=2.31.0
aiohttp>=3.9.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
